// src/types/store.ts
export interface storeData {
    name: string;
    url: string;
    logo: string;
    AffiliateUrl: string;
    description: string;
    AffiliateNetwork: string;
    AffiliateId: string;
    ApiKey: string;
    categories: string[];
    activated: boolean;
    priority: number;
    CashbackRate: string;
    StoreContactPersonName: string;
    StorePersonEmail: string;
    StorePersonPhone: string;
    storeBannerImage: string;
    SEOTitle: string;
    MetaDescription: string;
    revenue: number;
    clicks: number;
    conversionRate: number;
    webhookEndpoint: string;
  }